requirements.txt
Jinja2==2.7.2
boto3==1.1.4
